strip(tier-3): delete crates/firebase
The firebase crate was 145 lines of API-response schema types for the
Firebase auth REST API. Only two of those types were used outside the
crate (FirebaseError, FetchAccessTokenResponse) - both in
app/src/server/server_api/auth.rs (the Firebase token-refresh flow)
and one test file.

Inlined those two types into a new app/src/auth/firebase_types.rs
module and updated the imports. AccountInfo and GetAccountInfoResponse
were exported but unused; deleted with the crate.

The actual Firebase API call only fires when refreshing a real user's
auth tokens; with skip_login enabled the app uses Test credentials and
never reaches that code path. The http_client warp.dev block does not
cover Firebase domains anyway, so this is a meaningful cleanup at the
dependency layer.
