--- crates/firebase/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "firebase"
-authors = ["Warp Team <dev@warp.dev>"]
-version = "0.1.0"
-edition = "2021"
-description = "Utilities for firebase client APIs"
-publish.workspace = true
-license.workspace = true
-
-[dependencies]
-anyhow.workspace = true
-serde.workspace = true
